Implement #419: Add group column for motion csv import
The motion import now features an additional column to indicate that the import should be created by an existing group.
If the column is missing it is considered as 'off' and first / last name are used as always.
If the column exists and contains 'y','yes','j','ja','true', '1' or 1 first name will be ignored and last name will be interpreted as a group name.
If the group exists and can act as a person the motion will be imported and assigned to this group.

Signed-off-by: René Köcher <shirk@bitspin.org>