server: Fix database config in Docker entrypoint
Firstly, this patch ensures that the server connects to the database
according to the given configuration variables.  Up until now, there was
a pg_isready check for the hard-coded hostname "db".  While this
generally worked in the default setup – db is an alias for pgbouncer in
the provided YAML configuration files – it is obviously wrong and would
lead to unexpected behavior in customized setups.

Secondly, the .pgpass setup, which was a remnant of another time and not
used anymore, has been removed.