Enable pytest in docker
The DJANGO_SETTINGS_MODULE is set in the develop container, so the
tests.settings from the setup.cfg was ignored. Executing pytest with

    DJANGO_SETTINGS_MODULE=tests.settings pytest

works fine. The line added in the setup.cfg takes over precedence: It
auto-adds the --ds parameter which has a higher precedence than the
environment variable.

Also removed an unnecessary setting.